create `useSafeWebsite` hook
